+++ watch/utils/simple_dvc.py
@@ -34,6 +34,12 @@ class SimpleDVC(ub.NiceRepr):
     def __nice__(self):
         return str(self.dvc_root)
 
+    @property
+    def cache_dir(self):
+        info = ub.cmd('dvc cache dir', cwd=self.dvc_root, check=True)
+        cache_dpath = ub.Path(info['out'].strip())
+        return cache_dpath
+
     @classmethod
     def demo_dpath(cls, reset=False):
         dvc_dpath = ub.Path.appdir('simple_dvc/test/test_dvc_repo')